return true if no categories

+++ lib.php
@@ -81,6 +81,9 @@ function qpractice_add_instance(stdClass $qpractice, mod_qpractice_mod_form $mfo
 
 function upsert_categories(stdClass $qpractice){
     global $DB;
+    if(empty($qpractice->categories)){
+        return true;
+    }
     $DB->delete_records('qpractice_categories', ['qpracticeid' => $qpractice->id]);
     $recordstoinsert = [];
     foreach(array_keys($qpractice->categories) as $categoryid){
